pyproject.toml
[build-system]
requires = ["chronver", "setuptools>=66"]
build-backend = "setuptools.build_meta"

[project]
authors = [
    { name = "Robert Nagler", email = "rnpix@robnagler.com" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Topic :: Utilities",
]
dependencies = [
    "exif",
    "pillow",
    "pykern",
]
description = "Photo library tools"
dynamic = ["version"]
name = "rnpix"
readme = "README.md"

[project.scripts]
rnpix = "rnpix.rnpix_console:main"

[project.urls]
Homepage = "http://github.com/robnagler/rnpix"

[tool.setuptools.package-data]
rnpix = ["package_data/**"]

[tool.setuptools.packages.find]
include = ["rnpix*"]
